renamed userDataController file to controller
